+++ CHANGELOG.md
@@ -10,6 +10,8 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Adds CLI command to update individual alignments path
 
 ### Fixed
+- Variant loading assumes GATK if no caller set given and GATK filter status is seen in FILTER
+
 ### Changed